Create arduino_switch_dxlevent.py
This is POC code to test using IOT devices to add to security detection capabilities using OpenDXL as a messaging bus to send a message, and an Arduino with a Grove Shield with a switch to create an alert when the changes to 'open'. In testing, this switch is connected to a datacenter cabinet. When the cabinet is open, the button/switch state changes to open, sending an alert.